+++ R/tax_range_time.R
@@ -45,7 +45,7 @@
 #' # Grab internal data
 #' occdf <- tetrapods
 #' # Remove NAs
-#' occdf <- subset(occdf, !is.na(order))
+#' occdf <- subset(occdf, !is.na(order) & order != "NO_ORDER_SPECIFIED")
 #' # Temporal range
 #' ex <- tax_range_time(occdf = occdf, name = "order", plot = TRUE)
 #'