+++ NEWS.md
@@ -1,4 +1,9 @@
 # palaeoverse development version
+Updates
+
+* Added package-level documentation function
+* bin_time now accepts user-defined probability weights
+
 Minor bug fixes
 
 * Updated palaeorotate reconstruction files to use an hexagonal equal-area grid.
@@ -10,8 +15,6 @@ Minor bug fixes
 * Removed dependence of tests on divDyn and deeptime
 * Fixed tax_range_time example (#60)
 * Fixed look_up issue when handling pre-Cambrian occurrences.
-
-Added package-level documentation function
 
 # palaeoverse 1.0.0
 First full release of the palaeoverse R package.