+++ NEWS.md
@@ -5,6 +5,7 @@ Updates
 * lat_bins now accepts user-defined latitudinal range (#23)
 * bin_lat now has functionality for handling boundary occurrences (#23)
 * tax_unique now supports arbitrary higher taxonomic levels (#52)
+* time_bins now supports user input dataframes (#19)
 
 Minor bug fixes